SecurityManager tests divided.
